+++ packages/relay-bitcoin-wallet-adapter/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @reservoir0x/relay-bitcoin-wallet-adapter
 
+## 17.0.2
+
+### Patch Changes
+
+- daaae5b: bump bitcoinjs-lib dep version
+
+## 17.0.1
+
+### Patch Changes
+
+- Updated dependencies [7b351ab]
+  - @relayprotocol/relay-sdk@5.2.1
+
+## 17.0.0
+
+### Patch Changes
+
+- Updated dependencies [144f7a4]
+  - @relayprotocol/relay-sdk@5.2.0
+
 ## 16.0.1
 
 ### Patch Changes